+++ src/ts/examples/tutorials/ex8.c
@@ -334,7 +334,7 @@ static PetscErrorCode MonitorError(TS ts,PetscInt step,PetscReal t,Vec x,void *c
   ierr = VecAYPX(mon->x,-1,x);CHKERRQ(ierr);
   ierr = VecNorm(mon->x,NORM_2,&nrm_diff);CHKERRQ(ierr);
   ierr = TSGetTimeStep(ts,&h);CHKERRQ(ierr);
-  ierr = PetscPrintf(mon->comm,"step %4D t=%12.8e h=% 8.2e  |x|=%9.2e  |x_e|=%9.2e  |x-x_e|=%9.2e\n",step,t,h,nrm_x,nrm_exact,nrm_diff);CHKERRQ(ierr);
+  ierr = PetscPrintf(mon->comm,"step %4D t=%12.8e h=% 8.2e  |x|=%9.2e  |x_e|=%9.2e  |x-x_e|=%9.2e\n",step,(double)t,(double)h,(double)nrm_x,(double)nrm_exact,(double)nrm_diff);CHKERRQ(ierr);
   PetscFunctionReturn(0);
 }
